warnings fix?

+++ Sources/QuickJSC/include/quickjs.h
@@ -523,7 +523,7 @@ static js_force_inline JSValue JS_NewInt64(JSContext *ctx, int64_t val)
 {
     JSValue v;
     if (val == (int32_t)val) {
-        v = JS_NewInt32(ctx, val);
+        v = JS_NewInt32(ctx, (int32_t)val);
     } else {
         v = __JS_NewFloat64(ctx, val);
     }
@@ -606,12 +606,12 @@ static inline JS_BOOL JS_IsUndefined(JSValueConst v)
 
 static inline JS_BOOL JS_IsException(JSValueConst v)
 {
-    return js_unlikely(JS_VALUE_GET_TAG(v) == JS_TAG_EXCEPTION);
+    return js_unlikely(JS_VALUE_GET_TAG(v) == (int)JS_TAG_EXCEPTION);
 }
 
 static inline JS_BOOL JS_IsUninitialized(JSValueConst v)
 {
-    return js_unlikely(JS_VALUE_GET_TAG(v) == JS_TAG_UNINITIALIZED);
+    return js_unlikely(JS_VALUE_GET_TAG(v) == (int)JS_TAG_UNINITIALIZED);
 }
 
 static inline JS_BOOL JS_IsString(JSValueConst v)
